+++ src/ewatercycle/models/__init__.py
@@ -1,3 +1,4 @@
+from .hype import Hype
 from .lisflood import Lisflood
 from .marrmot import MarrmotM01, MarrmotM14
 from .pcrglobwb import PCRGlobWB